:truck: [client] Move mockBackend to client/test
